+++ src/turn/tests/turnclienttest/tcpinitiator.h
@@ -128,35 +128,6 @@ struct TCPInitiator : public turn::TCPClientObserver
 
     void onRelayDataReceived(turn::Client&, const char* data, size_t size, const net::Address& peerAddr)
     {
-#if 0
-        std::string payload(data, size);
-        payload.erase(std::remove(payload.begin(), payload.end(), 'x'),
-        payload.end());
-        if (payload.length() == 8) {
-            uint64_t sentAt = util::strtoi<uint64_t>(payload);
-            uint64_t latency = time::ticks() - sentAt;
-
-            SDebug << id << ": Received data from " << peerAddr 
-                << ": payload=" << payload << ", latency=" << latency << endl;
-        }
-
-        if (size < 150) {
-            //std::string payload(data, size);
-            std::string payload(data, 8); // read the first packet from joined packets
-            uint64_t sentAt = util::strtoi<uint64_t>(payload);
-            uint64_t latency = time::ticks() - sentAt;
-
-            SDebug << id << ": Received data from " << peerAddr 
-                << ": payload=" << payload << ", latency=" << latency << endl;
-        }
-        else
-            LDebug(id << ": Received dummy data from " << peerAddr << ": size=", size);
-
-        // Echo back to peer
-        LDebug(id << ": Received data from  " << peerAddr << ": ", size);
-
-        client.sendData(data, size, peerAddr);
-#endif
         LDebug(id, ": Data received from responder");
 
         sendPacketToResponder();